# src/main/java/com/artcode/quarkus/user/LoginRequest.java
package com.artcode.quarkus.user;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class LoginRequest {

    public String email;
    public String password;
}