# src/main/java/com/artcode/quarkus/user/UserRequest.java
package com.artcode.quarkus.user;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class UserRequest {

    public String name;
    public String email;
}